requirements.txt
# Core Dependencies
numpy
pandas
matplotlib
seaborn
scikit-learn
ta  # Technical Analysis Library
pandas-ta # Another Technical Analysis Library

# Machine Learning
xgboost

# Data & Finance
yfinance
pandas_datareader

# Visualization
plotly
streamlit

# Reinforcement Learning & Trading
gymnasium
stable-baselines3[extra]  # Includes extra dependencies like OpenAI Gym wrappers
sb3-contrib  # Additional algorithms for Stable-Baselines3

# Environment-Specific
git+https://github.com/DLR-RM/stable-baselines3.git  # Latest stable-baselines3 updates

# Logging
logging  # Python's built-in logging, no need to install separately

# YAML Parsing
PyYAML
